removed unused stuff, renamed utilities to helper
