+++ packages/asset-kit/src/helpers/index.ts
@@ -1,3 +1,3 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
-export * from './helpers';
+export * from './AssetKitHelpers';
